simple changes

+++ codes/mcmc/test_mymcmc/prepare_mock_jackknife.py
@@ -38,7 +38,7 @@ def main():
         N = len(star_list) / N_jackknife
 
         for i in range(N_jackknife):
-            output_list = star_list[:(N * i)] + star_list[(N * i + N):]
+            output_list = star_list[:int(N * i)] + star_list[int(N * i + N):]
 
             output_filename = jk_dir + 'mock_' + p.ID + '_jk_' + str(i) + '.dat'
             output_file = open(output_filename, 'w')
